requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=6.0
wandb>=0.13.0
tensorboard>=2.10.0
datasets>=2.0.0
tokenizers>=0.13.0
transformers>=4.20.0
scikit-learn>=1.1.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
pre-commit>=2.20.0 